requirements.txt
python-kasa==0.7.0.3